Add logic to run version specific test suite
After merging the required changes to wso2/testgrid-job-configs to
send current product version, this logic should be updated to exit
if 'ProductVersion' is not found or empty. Can't run default test
suite when 'ProductVersion' is empty because it can later lead to
confusions on which test suite ran for a product version.