Remove use of root parent declaration
This is causing issues when running in test grid. When parent is
defined, maven build will create a /target after the build. TG
detects this /target as a scenario and tries to fetch details and
reports from this dir. When the required files are not available
build is marked as an failure.