Add a test scenario to test API creation with only sandpoint endpoint
This commit is to add a new test scenario to test API creation where it doesn’t have a product endpoint.
In addition to that, another test was added to test the key generation in that scenario.
Test scenario was added to APIM18CreateAnAPIThroughThePublisherRestAPITestCase.
This test will cover https://github.com/wso2/api-manager/issues/1190 fix.